requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
langgraph>=0.4.0
langchain-groq>=0.3.0
langchain-core>=0.3.0
python-dotenv>=1.1.0
pymongo>=4.12.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
